added explicist conversion from int? to int in edit quesitons page view model

+++ Trivia_Stage2/ViewModels/EditQuestionsPageViewModel.cs
@@ -20,14 +20,14 @@ namespace Trivia_Stage2.ViewModels
             if (Question != null)
             {
                 QuestionId = Question.QuestionId;
-                PlayerId = Question.PlayerId;
+                PlayerId = (int)Question.PlayerId;
                 Correct = Question.Correct;
                 Incorrect1 = Question.Incorrect1;
                 Incorrect2 = Question.Incorrect2;
                 Incorrect3 = Question.Incorrect3;
                 QuestionText = Question.QuestionText;
-                SubjectId = Question.SubjectId;
-                StatusId = Question.StatusId;
+                SubjectId = (int)Question.SubjectId;
+                StatusId = (int)Question.StatusId;
                 Player = Question.Player;
                 Status = Question.Status;
                 Subject= Question.Subject;
